apply finalized changes
